+++ frontend/src/pages/EditFacturaPage/FormaPagoModal.tsx
@@ -14,20 +14,20 @@ import {
     useTheme
 } from "@mui/material";
 
-import {type ImpuestoType} from "../factura";
-import DialogTitle from "../../../components/ui/DialogTitle";
-import { StyledInput } from '../../../components/ui/StyledInput/StyledInput';
-import {TARIFAS_IVA} from "../../tarifario";
+import {type PagoType} from "../FacturasPage/factura";
+import DialogTitle from "../../components/ui/DialogTitle";
+import { StyledInput } from '../../components/ui/StyledInput/StyledInput.tsx';
+import {FORMAS_PAGO} from "../tarifario.ts";
 
-interface EditImpuestosProps {
-    impuesto?: ImpuestoType;
-    onUpdate?: (impuestos: ImpuestoType) => void;
+interface FormaPagoModalProps {
+    pago?: PagoType;
+    onUpdate?: (pago: PagoType) => void;
     isVisible: boolean;
     setVisible: (visible: boolean) => void;
 }
 
-const EditImpuestosModal: React.FC<EditImpuestosProps> = ({
-    impuesto,
+const FormaPagoModalModal: React.FC<FormaPagoModalProps> = ({
+    pago,
     onUpdate,
     isVisible,
     setVisible
@@ -37,67 +37,67 @@ const EditImpuestosModal: React.FC<EditImpuestosProps> = ({
     reset,
     handleSubmit,
     formState: { errors }
-  } = useForm<ImpuestoType>({
+  } = useForm<PagoType>({
     defaultValues: {
-      codigo: 2,
-      codigoPorcentaje: 2,
-      tarifa: 12,
-      baseImponible: 0,
-      valor: 0
+      formaPago: '01',
+      total: 0,
+      plazo: 0,
+      unidadTiempo: 'días'
     }
   });
   const formRef = useRef<HTMLFormElement>(null);
   const theme = useTheme();
 
   useEffect(() => {
-    reset(impuesto ?? {codigo: 2, codigoPorcentaje: 2, tarifa: 12, baseImponible: 0, valor: 0});
-  }, [impuesto, reset]);
+    reset(pago ?? {formaPago: '01', total: 0, plazo: 0, unidadTiempo: 'días'});
+  }, [pago, reset]);
 
   const handleCloseDialog = () => setVisible(false);
 
   const onFormSubmit = handleSubmit((data) => {
-    const index = TARIFAS_IVA
-      .findIndex(({ codigoPorcentaje }) => codigoPorcentaje === +data.codigoPorcentaje);
-    const impuesto = {
-      ...data,
-      codigoPorcentaje: +data.codigoPorcentaje,
-      tarifa: TARIFAS_IVA[index]?.tarifa
-    } as ImpuestoType;
-    onUpdate && onUpdate(impuesto);
+    // const index = FORMAS_PAGO
+    //   .findIndex(({ codigo }) => codigo === data.formaPago);
+    const pago = {
+      formaPago: data.formaPago,
+      total: data.total,
+      plazo: data.plazo,
+      unidadTiempo: 'días'
+    } as PagoType;
+    onUpdate && onUpdate(pago);
   })
   return (
     <Dialog
       fullWidth
       maxWidth="sm"
       onClose={handleCloseDialog}
-      aria-labelledby="edit-impuesto-title"
+      aria-labelledby="edit-forma-pago-title"
       open={isVisible}
     >
-      <DialogTitle id="edit-impuesto-title" onClose={handleCloseDialog}>
-        Seleccione el Impuesto a Aplicar
+      <DialogTitle id="edit-forma-pago-title" onClose={handleCloseDialog}>
+        Seleccione forma de pago
       </DialogTitle>
       <DialogContent sx={[ { minHeight: '240px', }, ]}>
         <form ref={formRef} noValidate autoComplete="off" onSubmit={onFormSubmit}>
-          <FormControl fullWidth variant="outlined" error={errors.codigoPorcentaje !== undefined}
+          <FormControl fullWidth variant="outlined" error={errors.formaPago !== undefined}
              css={css`.Mui-error>select { border-color: ${theme.palette.error.main} !important; }`}>
             <InputLabel
                 htmlFor="device-type-select"
                 shrink={false}
                 css={css`margin-bottom: -24px; transform: none; position: relative; color: ${theme.palette.text.primary}!important; `}
             >
-                Tarifa del IVA *
+                Forma de Pago *
             </InputLabel>
             <NativeSelect
               id="device-type-select"
               input={<StyledInput />}
-              {...register("codigoPorcentaje", { required: true })}
+              {...register("formaPago", { required: true })}
             >
-              <option value="" disabled>Tarifa del IVA</option>
-              {TARIFAS_IVA.filter(({active}) => active).map(
-                tarifa => <option key={tarifa.codigoPorcentaje} value={tarifa.codigoPorcentaje}>{tarifa.label}</option>
+              <option value="" disabled>Forma de Pago</option>
+              {FORMAS_PAGO.map(
+                forma => <option key={forma.codigo} value={forma.codigo}>{forma.label}</option>
               )}
             </NativeSelect>
-            <FormHelperText>{errors?.codigoPorcentaje?.message}</FormHelperText>
+            <FormHelperText>{errors?.formaPago?.message}</FormHelperText>
           </FormControl>
         </form>
       </DialogContent>
@@ -112,4 +112,4 @@ const EditImpuestosModal: React.FC<EditImpuestosProps> = ({
   );
 }
 
-export default EditImpuestosModal;
+export default FormaPagoModalModal;
