+++ frontend/src/pages/EditFacturaPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './EditFacturaPage';
