+++ frontend/src/pages/FacturasPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FacturasPage';
