+++ frontend/src/services/auth/withAuth.tsx
@@ -1,6 +1,6 @@
 import RequireAuth from "./RequireAuth.tsx";
 
-const withAuth = (Component: React.ComponentType) => (props: object) => {
+const withAuth = (Component: React.ComponentType | React.FC<any>) => (props: object) => {
   return <RequireAuth><Component {...props} /></RequireAuth>;
 };
